Move client dataset services to api/datasets
